refactor : useLocalStorage 추가, 테스트 코드 분리
