+++ src/refactoring/App.tsx
@@ -50,6 +50,7 @@ const initialCoupons: Coupon[] = [
 const App = () => {
   const { products, updateProduct, addProduct } = useProducts(initialProducts);
   const { coupons, addCoupon } = useCoupons(initialCoupons);
+
   const [isAdmin, setIsAdmin] = useState(false);
 
   const handleToggleAdmin = () => {
